DOTFILES_DIR -> STOW_DIR

+++ up.sh
@@ -6,34 +6,33 @@ if [ -d $HOME/.zprezto/.git ]; then
   ( git pull && git submodule update --init --recursive ) &> /dev/null
   echo "done!"
 else
-  if [ ! -d $HOME/.zprezto ]; then
-    echo "Installing Prezto...\c"
-    ( git clone https://github.com/zsh-users/prezto.git $HOME/.zprezto && ( cd $HOME/.zprezto && git submodule update --init --recursive ) ) &> /dev/null
-    echo "done!"
-  fi
+  echo "Installing Prezto...\c"
+  [ -d $HOME/.zprezto ] && rm -rf $HOME/.zprezto
+  ( git clone https://github.com/zsh-users/prezto.git $HOME/.zprezto && ( cd $HOME/.zprezto && git submodule update --init --recursive ) ) &> /dev/null
+  echo "done!"
 fi
 
-if [ -d $HOME/.config/base16-shell/.git ]; then
+if [ -d $HOME/.config/base16-shell ]; then
   echo "Removing Base16 Shell...\c"
   rm -rf $HOME/.config/base16-shell
   echo "done!"
 fi
 
 if [ ! -d $HOME/.vim/bundle/repos/github.com/Shougo/dein.vim ]; then
-  echo "Uninstalling NeoBundle...\c"
-  rm -rf $HOME/.vim/bundle
-  echo "done!"
   echo "Installing Dein.vim...\c"
-  mkdir -p $HOME/.vim/bundle/repos/github.com/Shougo
+  [ -d $HOME/.vim/bundle ] && rm -rf $HOME/.vim/bundle
   ( git clone https://github.com/Shougo/dein.vim $HOME/.vim/bundle/repos/github.com/Shougo/dein.vim ) &> /dev/null
   echo "done!"
 fi
 
-echo "Updating Vim plugins...\c"
-$VISUAL -u $HOME/.vimrc.bundles -e -s "+call dein#update()" +qa
-echo "done!"
+if [[ $STOW_DIR ]]; then
+  echo "Stowing common dotfiles...\c"
+  stow -t $HOME -R common
+  echo "done!"
+fi
 
-echo "Pruning removed dotfiles...\c"
+echo "Pruning broken dotfile links...\c"
 for dotted in .*; do
   find -L $dotted -type l | xargs rm
 done
+echo "done!"
